+++ src/proximal_maps.jl
@@ -2,13 +2,13 @@
     y = prox_distance(M, λ, f, p [, r=2])
     prox_distance!(M, q, λ, f, p [, r=2])
 
-compute the proximal map ``\operatorname{prox}_{λφ}`` with
+Compute the proximal map ``\operatorname{prox}_{λφ}`` with
 parameter λ of ``φ(p) = \frac{1}{r}d_{\mathcal M}^r(f,p)``.
 For the in-place variant the computation is done in place of `q`.
 
 # Input
 * `M` – a manifold `M`
-* `λ` – the prox parameter
+* `λ` – the prox parameter, a positive real number.
 * `f` – a point ``f ∈ \mathcal M`` (the data)
 * `p` – the argument of the proximal map
 * `r` – (`2`) exponent of the distance.
@@ -18,7 +18,7 @@ For the in-place variant the computation is done in place of `q`.
 
 For more details see [WeinmannDemaretStorath:2014](@cite)
 """
-function prox_distance(M::AbstractManifold, λ, f, p, r::Int = 2)
+function prox_distance(M::AbstractManifold, λ::Real, f, p, r::Int = 2)
     d = distance(M, f, p)
     if r == 2
         t = λ / (1 + λ)
@@ -31,7 +31,7 @@ function prox_distance(M::AbstractManifold, λ, f, p, r::Int = 2)
             ),
         )
     end
-    return exp(M, p, log(M, p, f), t)
+    return shortest_geodesic(M, p, f, t)
 end
 function prox_distance!(M::AbstractManifold, q, λ, f, p, r::Int = 2)
     d = distance(M, f, p)
@@ -46,5 +46,5 @@ function prox_distance!(M::AbstractManifold, q, λ, f, p, r::Int = 2)
             ),
         )
     end
-    return exp!(M, q, p, log(M, p, f), t)
+    return shortest_geodesic!(M, q, p, f, t)
 end